+++ src/compile/reductions.c
@@ -22,7 +22,7 @@ Text_t compile_reduction(env_t *env, ast_t *ast) {
     static int64_t next_id = 1;
     ast_t *item = FakeAST(Var, String("$it", next_id++));
     ast_t *body = LiteralCode(Text("{}")); // placeholder
-    ast_t *loop = FakeAST(For, .vars = new (ast_list_t, .ast = item), .iter = reduction->iter, .body = body);
+    ast_t *loop = FakeAST(For, .vars = new (ast_list_t, .ast = item), .iters = new (ast_list_t, .ast = reduction->iter), .body = body);
     env_t *body_scope = for_scope(env, loop);
     if (op == Equals || op == NotEquals || op == LessThan || op == LessThanOrEquals || op == GreaterThan
         || op == GreaterThanOrEquals) {