+++ src/formatter/formatter.c
@@ -209,7 +209,12 @@ OptionalText_t format_inline_code(ast_t *ast, Table_t comments) {
             if (var->next) code = Texts(code, ", ");
         }
         if (loop->at) code = Texts(code, " at ", fmt_inline(loop->at, comments));
-        code = Texts(code, " in ", fmt_inline(loop->iter, comments), " do ", fmt_inline(loop->body, comments));
+        code = Texts(code, " in ");
+        for (ast_list_t *iter = loop->iters; iter; iter = iter->next) {
+            code = Texts(code, fmt_inline(iter->ast, comments));
+            if (iter->next) code = Texts(code, ", ");
+        }
+        code = Texts(code, " do ", fmt_inline(loop->body, comments));
         if (loop->empty) code = Texts(code, " else ", fmt_inline(loop->empty, comments));
         return code;
     }
@@ -221,7 +226,11 @@ OptionalText_t format_inline_code(ast_t *ast, Table_t comments) {
             if (var->next) code = Texts(code, ", ");
         }
         if (comp->at) code = Texts(code, " at ", fmt_inline(comp->at, comments));
-        code = Texts(code, " in ", fmt_inline(comp->iter, comments));
+        code = Texts(code, " in ");
+        for (ast_list_t *iter = comp->iters; iter; iter = iter->next) {
+            code = Texts(code, fmt_inline(iter->ast, comments));
+            if (iter->next) code = Texts(code, ", ");
+        }
         if (comp->filter) code = Texts(code, " if ", fmt_inline(comp->filter, comments));
         return code;
     }
@@ -346,10 +355,10 @@ OptionalText_t format_inline_code(ast_t *ast, Table_t comments) {
     /*inline*/ case Reduction: {
         DeclareMatch(reduction, ast, Reduction);
         if (reduction->key) {
-            return Texts("(", fmt_inline(reduction->key, comments), ": ", fmt_inline(reduction->iter, comments));
+            return Texts("(", fmt_inline(reduction->key, comments), ": ", fmt_inline(reduction->iter, comments), ")");
         } else {
             return Texts("(", Text$from_str(binop_info[reduction->op].operator), ": ",
-                         fmt_inline(reduction->iter, comments));
+                         fmt_inline(reduction->iter, comments), ")");
         }
     }
     /*inline*/ case None:
@@ -542,7 +551,12 @@ Text_t format_code(ast_t *ast, Table_t comments, Text_t indent) {
             if (var->next) code = Texts(code, ", ");
         }
         if (loop->at) code = Texts(code, " at ", fmt(loop->at, comments, indent));
-        code = Texts(code, " in ", fmt(loop->iter, comments, indent), format_namespace(loop->body, comments, indent));
+        code = Texts(code, " in ");
+        for (ast_list_t *iter = loop->iters; iter; iter = iter->next) {
+            code = Texts(code, fmt(iter->ast, comments, indent));
+            if (iter->next) code = Texts(code, ", ");
+        }
+        code = Texts(code, format_namespace(loop->body, comments, indent));
         if (loop->empty) code = Texts(code, "\n", indent, "else", format_namespace(loop->empty, comments, indent));
         return code;
     }
@@ -559,7 +573,11 @@ Text_t format_code(ast_t *ast, Table_t comments, Text_t indent) {
         }
         if (comp->at) code = Texts(code, " at ", fmt(comp->at, comments, indent));
 
-        code = Texts(code, " in ", fmt(comp->iter, comments, indent));
+        code = Texts(code, " in ");
+        for (ast_list_t *iter = comp->iters; iter; iter = iter->next) {
+            code = Texts(code, fmt(iter->ast, comments, indent));
+            if (iter->next) code = Texts(code, ", ");
+        }
 
         if (comp->filter) {
             if (code.length >= MAX_WIDTH) code = Texts(code, "\n", indent, "if ");
@@ -812,10 +830,10 @@ Text_t format_code(ast_t *ast, Table_t comments, Text_t indent) {
         DeclareMatch(reduction, ast, Reduction);
         if (reduction->key) {
             return Texts("(", fmt(reduction->key, comments, Texts(indent, single_indent)), ": ",
-                         fmt(reduction->iter, comments, Texts(indent, single_indent)));
+                         fmt(reduction->iter, comments, Texts(indent, single_indent)), ")");
         } else {
             return Texts("(", binop_info[reduction->op].operator, ": ",
-                         fmt(reduction->iter, comments, Texts(indent, single_indent)));
+                         fmt(reduction->iter, comments, Texts(indent, single_indent)), ")");
         }
     }
     /*multiline*/ case Stop: